Reorganize akd_ext.agents._base into package
- Split akd_ext/agents/_base.py and akd_ext/agents/_base_pydantic/
  into a single _base/ package with openai.py and pydantic_ai/
  submodule. External imports via akd_ext.agents._base stay intact.
- Relocate cmr_care_pydantic.py validation agent to project-root
  examples/ since it's not part of the shipped API.
- Rename docs/pydantic_ai_base_agent_diagram_full.{md,svg} to
  docs/pydantic_ai_base.{md,svg}.
- Pin mistralai<2 via tool.uv.override-dependencies to work around
  instructor 1.13 eagerly importing from mistralai 2.x.
- Add pythonpath=["."] to pytest config so tests can import examples/.
