+++ akd_ext/agents/_base/__init__.py
@@ -1,0 +1,17 @@
+"""Base agent classes for akd_ext.
+
+Two families live here:
+
+- :class:`OpenAIBaseAgent` — built on the OpenAI Agents SDK (``openai``).
+- :class:`PydanticAIBaseAgent` — built on :class:`pydantic_ai.Agent`.
+"""
+
+from akd_ext.agents._base.openai import OpenAIBaseAgent, OpenAIBaseAgentConfig
+from akd_ext.agents._base.pydantic_ai import PydanticAIBaseAgent, PydanticAIBaseAgentConfig
+
+__all__ = [
+    "OpenAIBaseAgent",
+    "OpenAIBaseAgentConfig",
+    "PydanticAIBaseAgent",
+    "PydanticAIBaseAgentConfig",
+]
