+++ akd_ext/agents/_base/pydantic_ai/_capabilities.py
