+++ pyproject.toml
@@ -52,6 +52,7 @@ exclude = ["tests*"]
 [tool.pytest.ini_options]
 asyncio_mode = "auto"
 testpaths = ["tests"]
+pythonpath = ["."]
 python_files = ["test_*.py", "*_test.py"]
 python_functions = ["test_*"]
 addopts = [
@@ -71,7 +72,7 @@ markers = [
 ]
 
 [tool.uv]
-override-dependencies = ["wrapt>=1.14"]
+override-dependencies = ["wrapt>=1.14", "mistralai<2"]
 
 [tool.ruff]
 target-version = "py312"
